✨ [FEAT] 특정 가게 리뷰 목록 조회 API

+++ src/main/java/umc/spring/domain/Review.java
@@ -19,7 +19,7 @@ public class Review extends BaseEntity {
     private Long id;
 
     @Column(nullable = false)
-    private Double ratePoint;
+    private Float ratePoint;
 
     @Column(nullable = false, columnDefinition = "TEXT")
     private String reviewText;
